pyproject.toml
[project]
name = "hinomaru-home-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cotton>=1.5.2",
    "django-tailwind-cli>=2.21.1",
    "django-widget-tweaks>=1.5.0",
    "faker>=34.0.2",
    "icecream>=2.1.4",
    "requests>=2.32.3",
    "wagtail>=6.3.2",
]
